Separating types, init states to own components
api and firebase files now together in services folder
